Support CKAN 2.10
