+++ ckanext/dcat/harvesters/rdf.py
@@ -14,11 +14,9 @@ import ckan.model as model
 import ckan.lib.plugins as lib_plugins
 
 from ckanext.harvest.model import HarvestObject, HarvestObjectExtra
-
+from ckanext.harvest.logic.schema import unicode_safe
 from ckanext.dcat.harvesters.base import DCATHarvester
-
 from ckanext.dcat.processors import RDFParserException, RDFParser
-
 from ckanext.dcat.interfaces import IDCATRDFHarvester
 
 
@@ -379,7 +377,7 @@ class DCATRDFHarvester(DCATHarvester):
 
                 # We need to explicitly provide a package ID
                 dataset['id'] = str(uuid.uuid4())
-                package_schema['id'] = [str]
+                package_schema['id'] = [unicode_safe]
 
                 harvester_tmp_dict = {}
