+++ ckanext/dcat/tests/test_blueprints.py
@@ -11,12 +11,12 @@ import pytest
 from ckan import plugins as p
 
 from rdflib import Graph
-from ckantoolkit import url_for as core_url_for
 from ckantoolkit.tests import factories
 
 from ckanext.dcat.processors import RDFParser
 from ckanext.dcat.profiles import RDF, DCAT
 from ckanext.dcat.processors import HYDRA
+from ckanext.dcat.urls import url_for
 
 
 def _sort_query_params(url):
@@ -29,28 +29,6 @@ def _sort_query_params(url):
         (parts.scheme, parts.netloc, parts.path, parts.params,
          encoded_qs, parts.fragment)
     )
-
-
-def url_for(*args, **kwargs):
-
-    if not p.toolkit.check_ckan_version(min_version='2.9'):
-
-        external = kwargs.pop('_external', False)
-        if external is not None:
-            kwargs['qualified'] = external
-
-        if len(args) and args[0] == 'dcat.read_dataset':
-            return core_url_for('dcat_dataset', **kwargs)
-        elif len(args) and args[0] == 'dcat.read_catalog':
-            return core_url_for('dcat_catalog', **kwargs)
-        elif len(args) and args[0] == 'dataset.new':
-            return core_url_for(controller='package', action='new', **kwargs)
-        elif len(args) and args[0] == 'dataset.read':
-            return core_url_for(controller='package', action='read', **kwargs)
-
-
-    return core_url_for(*args, **kwargs)
-
 
 @pytest.mark.usefixtures('with_plugins', 'clean_db', 'clean_index')
 class TestEndpoints():
